:truck: Rename: OutputHandler.js -> LottoPrinter.js
